fix(document): complete the schema for namedscore
