+++ docarray/array/mixins/__init__.py
@@ -20,6 +20,8 @@ from .plot import PlotMixin
 from .pydantic import PydanticMixin
 from .reduce import ReduceMixin
 from .sample import SampleMixin
+from .setitem import SetItemMixin
+from .delitem import DelItemMixin
 from .text import TextToolsMixin
 from .traverse import TraverseMixin
 
@@ -27,6 +29,8 @@ from .traverse import TraverseMixin
 class AllMixins(
     GetAttributeMixin,
     GetItemMixin,
+    SetItemMixin,
+    DelItemMixin,
     ContentPropertyMixin,
     PydanticMixin,
     GroupMixin,
