+++ docarray/array/mixins/delitem.py
@@ -1,0 +1,92 @@
+from abc import abstractmethod
+from typing import (
+    TYPE_CHECKING,
+    Sequence,
+)
+
+import numpy as np
+
+from ...helper import typename
+
+if TYPE_CHECKING:
+    from ...types import (
+        DocumentArrayIndexType,
+    )
+
+
+class DelItemMixin:
+    """Provide help function to enable advanced indexing in `__delitem__`"""
+
+    @abstractmethod
+    def _del_docs_by_mask(self, mask: Sequence[bool]):
+        ...
+
+    @abstractmethod
+    def _del_doc_by_offset(self, offset: int):
+        ...
+
+    @abstractmethod
+    def _del_doc_by_id(self, _id: str):
+        ...
+
+    @abstractmethod
+    def _del_docs_by_slice(self, _slice: slice):
+        ...
+
+    @abstractmethod
+    def _del_all_docs(self):
+        ...
+
+    def __delitem__(self, index: 'DocumentArrayIndexType'):
+        if isinstance(index, (int, np.generic)) and not isinstance(index, bool):
+            self._del_doc_by_offset(int(index))
+
+        elif isinstance(index, str):
+            if index.startswith('@'):
+                raise NotImplementedError(
+                    'Delete elements along traversal paths is not implemented'
+                )
+            else:
+                self._del_doc_by_id(index)
+        elif isinstance(index, slice):
+            self._del_docs_by_slice(index)
+        elif index is Ellipsis:
+            self._del_all_docs()
+        elif isinstance(index, Sequence):
+            if (
+                isinstance(index, tuple)
+                and len(index) == 2
+                and isinstance(index[0], (slice, Sequence))
+            ):
+                if isinstance(index[0], str) and isinstance(index[1], str):
+                    # ambiguity only comes from the second string
+                    if index[1] in self._id2offset:
+                        del self[index[0]]
+                        del self[index[1]]
+                    else:
+                        self._set_doc_attr_by_index(index[0], index[1], None)
+                elif isinstance(index[0], (slice, Sequence)):
+                    _attrs = index[1]
+                    if isinstance(_attrs, str):
+                        _attrs = (index[1],)
+                    for _d in self[index[0]]:
+                        for _aa in _attrs:
+                            self._set_doc_attr_by_index(_d.id, _aa, None)
+            elif isinstance(index[0], bool):
+                self._del_docs_by_mask(index)
+            elif isinstance(index[0], int):
+                for t in sorted(index, reverse=True):
+                    del self[t]
+            elif isinstance(index[0], str):
+                for t in index:
+                    del self[t]
+        elif isinstance(index, np.ndarray):
+            index = index.squeeze()
+            if index.ndim == 1:
+                del self[index.tolist()]
+            else:
+                raise IndexError(
+                    f'When using np.ndarray as index, its `ndim` must =1. However, receiving ndim={index.ndim}'
+                )
+        else:
+            raise IndexError(f'Unsupported index type {typename(index)}: {index}')
